+++ src/player/RLDifficultyTotalPopup.cpp
@@ -195,7 +195,6 @@ bool RLDifficultyTotalPopup::init() {
         return false;
     setTitle("Rated Layouts Classic: -");
     auto contentSize = m_mainLayer->getContentSize();
-    m_noElasticity = true;
     // spinner
     auto spinner = LoadingSpinner::create(36.f);
     spinner->setPosition({contentSize.width / 2.f, contentSize.height / 2.f});